Added MasterServerQuerier class to get DayZ servers from the Steam Master Servers since this provides a more reliable way to determine the actual port number to use to query the server information.

+++ DayZServerMonitorCore/MessageParser.cs
@@ -35,9 +35,9 @@ namespace DayZServerMonitorCore
             return BitConverter.ToInt16(GetBytes(2).ToArray(), 0);
         }
 
-        public short GetPort()
+        public ushort GetPort()
         {
-            return IPAddress.NetworkToHostOrder(GetShort());
+            return (ushort)IPAddress.NetworkToHostOrder(GetShort());
         }
 
         public IPAddress GetIPAddress()
